Update deploy.ps1
We've been trying to deploy the ARM template for a few days with no success.  We were finally able to get it to deploy by switching line 25 to the code included below.  After this change the deployment is working for us.  The issue appears to be related to the GetFileName .NET call.  Thanks for all that you do to supply REDCap to the community.  

$filename = Split-Path $zipUri -leaf